3619: Rename EntityAttributeTable files
