+++ common/src/View/EntityPropertyTable.h
